refactor(os-apps): split entity alias helpers
