+++ crates/temper-platform/src/os_apps/mod.rs
@@ -12,7 +12,6 @@ use std::time::Instant;
 
 use chrono::{DateTime, NaiveDateTime, Utc};
 use serde::Serialize;
-use sha2::{Digest, Sha256};
 use temper_runtime::tenant::TenantId;
 use temper_server::state::WasmModuleSource;
 use temper_spec::automaton;
@@ -24,6 +23,7 @@ use crate::state::PlatformState;
 mod agent_bootstrap;
 mod app_catalog;
 mod closure_bootstrap;
+mod entity_aliases;
 mod policy_rows;
 mod reconcile;
 mod runtime_heal;
@@ -38,6 +38,10 @@ pub use closure_bootstrap::{
     ClosureBootstrapResult, OS_APP_CLOSURE_RESOLVER_VERSION, OsAppClosure,
     bootstrap_closure_manifest, os_app_closure_for_roots, parse_bootstrap_manifest_str,
     startup_os_app_closure,
+};
+use entity_aliases::{
+    content_sha256, ensure_created_file_initialized, ensure_entity_field_aliases,
+    file_already_contains, slug_fragment,
 };
 #[cfg(test)]
 pub(super) use policy_rows::os_app_policy_row_id;
@@ -2527,127 +2531,6 @@ pub(super) async fn ensure_markdown_file(
         .map_err(|e| format!("failed to upload File('{}') content: {e}", target.file_id))?;
 
     Ok(())
-}
-
-async fn ensure_created_file_initialized(
-    state: &PlatformState,
-    tenant_id: &TenantId,
-    agent_ctx: &temper_server::request_context::AgentContext,
-    file_id: &str,
-    create_fields: serde_json::Value,
-) -> Result<(), String> {
-    let response = state
-        .server
-        .get_tenant_entity_state(tenant_id, "File", file_id)
-        .await
-        .map_err(|e| format!("failed to inspect File('{file_id}') status: {e}"))?;
-    if response.state.status != "Created" {
-        return Ok(());
-    }
-
-    state
-        .server
-        .dispatch(temper_server::state::DispatchCommand {
-            tenant: tenant_id,
-            entity_type: "File",
-            entity_id: file_id,
-            action: "Create",
-            params: create_fields,
-            agent_ctx,
-            await_integration: false,
-            await_reactions: true,
-        })
-        .await
-        .map(|_| ())
-        .map_err(|e| format!("failed to initialize File('{file_id}'): {e}"))
-}
-
-async fn ensure_entity_field_aliases(
-    state: &PlatformState,
-    tenant_id: &TenantId,
-    entity_type: &str,
-    entity_id: &str,
-    aliases: serde_json::Value,
-) -> Result<(), String> {
-    let current = state
-        .server
-        .get_tenant_entity_state(tenant_id, entity_type, entity_id)
-        .await
-        .map_err(|e| format!("failed to inspect {entity_type}('{entity_id}') aliases: {e}"))?;
-
-    let Some(alias_map) = aliases.as_object() else {
-        return Ok(());
-    };
-    let mut updates = serde_json::Map::new();
-    for (key, expected) in alias_map {
-        if current.state.fields.get(key) != Some(expected) {
-            updates.insert(key.clone(), expected.clone());
-        }
-    }
-    if updates.is_empty() {
-        return Ok(());
-    }
-
-    state
-        .server
-        .update_tenant_entity_fields(
-            tenant_id,
-            entity_type,
-            entity_id,
-            serde_json::Value::Object(updates),
-            false,
-        )
-        .await
-        .map(|_| ())
-        .map_err(|e| format!("failed to repair {entity_type}('{entity_id}') aliases: {e}"))
-}
-
-async fn file_already_contains(
-    state: &PlatformState,
-    tenant_id: &TenantId,
-    file_id: &str,
-    desired_hash: &str,
-) -> Result<bool, String> {
-    let response = state
-        .server
-        .get_tenant_entity_state(tenant_id, "File", file_id)
-        .await
-        .map_err(|e| format!("failed to inspect File('{file_id}'): {e}"))?;
-
-    let has_content = response
-        .state
-        .booleans
-        .get("has_content")
-        .copied()
-        .unwrap_or(false);
-    let current_hash = response
-        .state
-        .fields
-        .get("content_hash")
-        .and_then(|value| value.as_str())
-        .unwrap_or_default();
-    Ok(has_content && current_hash == desired_hash)
-}
-
-fn content_sha256(content: &[u8]) -> String {
-    let mut hasher = Sha256::new();
-    hasher.update(content);
-    format!("sha256:{:x}", hasher.finalize())
-}
-
-pub(super) fn slug_fragment(value: &str) -> String {
-    let mut slug = String::new();
-    let mut last_was_sep = true;
-    for ch in value.chars() {
-        if ch.is_ascii_alphanumeric() {
-            slug.push(ch.to_ascii_lowercase());
-            last_was_sep = false;
-        } else if !last_was_sep {
-            slug.push('-');
-            last_was_sep = true;
-        }
-    }
-    slug.trim_matches('-').to_string()
 }
 
 /// Bootstrap seed data instances into the tenant.
